+++ compile-tsc
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -e
 
 rm -rf build/*
 mkdir -p build
