+++ src/external/index.ts
@@ -14,12 +14,7 @@
  * limitations under the License.
  */
 
-export * from './attributeInfo';
-export * from './common';
-export * from './dataset';
-export * from './external';
-export * from './set';
-export * from './range';
-export * from './numberRange';
-export * from './timeRange';
-export * from './stringRange';
+export * from './sqlExternal';
+export * from './mySqlExternal';
+export * from './postgresExternal';
+export * from './druidExternal';
