+++ src/helper/utils.ts
@@ -14,8 +14,6 @@
  * limitations under the License.
  */
 
-import { Expression, ExpressionJS } from "../expressions/baseExpression";
-
 var objectHasOwnProperty = Object.prototype.hasOwnProperty;
 export function hasOwnProperty(obj: any, key: string): boolean {
   return objectHasOwnProperty.call(obj, key);
@@ -159,24 +157,6 @@ export function emptyLookup(lookup: Lookup<any>): boolean {
 
 export function nonEmptyLookup(lookup: Lookup<any>): boolean {
   return !emptyLookup(lookup);
-}
-
-export function expressionLookupFromJS(expressionJSs: Lookup<ExpressionJS>): Lookup<Expression> {
-  var expressions: Lookup<Expression> = Object.create(null);
-  for (var name in expressionJSs) {
-    if (!hasOwnProperty(expressionJSs, name)) continue;
-    expressions[name] = Expression.fromJSLoose(expressionJSs[name]);
-  }
-  return expressions;
-}
-
-export function expressionLookupToJS(expressions: Lookup<Expression>): Lookup<ExpressionJS> {
-  var expressionsJSs: Lookup<ExpressionJS> = {};
-  for (var name in expressions) {
-    if (!hasOwnProperty(expressions, name)) continue;
-    expressionsJSs[name] = expressions[name].toJS();
-  }
-  return expressionsJSs;
 }
 
 export function safeAdd(num: number, delta: number): number {
