+++ src/index.ts
@@ -15,6 +15,12 @@
  */
 
 export const version = '###_VERSION_###';
+
+export * from './helper/index';
+export * from './dialect/index';
+export * from './datatypes/index';
+export * from './external/index';
 export * from './expressions/index';
 export * from './actions/index';
-export * from './datatypes/index';
+export * from './executor/index';
+
